refactor(tui): move app tests into the submodules they exercise
